Add security

+++ src/main/java/com/stocks/project/repository/SecurityRepository.java
@@ -1,8 +1,15 @@
 package com.stocks.project.repository;
 
+import com.stocks.project.exception.EmailOrUsernameIsAlreadyUsedException;
+import com.stocks.project.exception.NoSuchUserException;
+import com.stocks.project.exception.NotEnoughDataException;
+import com.stocks.project.model.Role;
 import com.stocks.project.model.SecurityInfo;
+import com.stocks.project.model.UserSecurityDTO;
 import com.stocks.project.utils.SecurityMapper;
-import org.springframework.beans.factory.annotation.Autowired;
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Repository;
 
 import javax.sql.DataSource;
@@ -13,25 +20,22 @@ import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Repository
+@RequiredArgsConstructor
 public class SecurityRepository {
     private final DataSource dataSource;
     private final SecurityMapper securityMapper;
-
-    @Autowired
-    public SecurityRepository(DataSource dataSource, SecurityMapper securityMapper) {
-        this.dataSource = dataSource;
-        this.securityMapper = securityMapper;
-    }
+    private final UserRepository userRepository;
+    private final PasswordEncoder passwordEncoder;
 
     public List<SecurityInfo> findAll() {
         List<SecurityInfo> securityInfos = new ArrayList<>();
-        String query = "SELECT * FROM security_info;";
-        try(
-                Connection connection = dataSource.getConnection();
-                PreparedStatement preparedStatement = connection.prepareStatement(query)
-        ) {
+        String query = "SELECT * FROM security_info INNER JOIN public.role r " +
+                "ON r.role_id = security_info.role_id;";
+        try (Connection connection = dataSource.getConnection();
+             PreparedStatement preparedStatement = connection.prepareStatement(query)) {
             ResultSet resultSet = preparedStatement.executeQuery();
             while (resultSet.next()) {
                 securityInfos.add(securityMapper.mapRow(resultSet));
@@ -43,13 +47,12 @@ public class SecurityRepository {
         return securityInfos;
     }
 
-    public SecurityInfo findById(int id) {
+    public Optional<SecurityInfo> findById(int id) {
         SecurityInfo securityInfo = null;
-        String query = "SELECT * FROM security_info WHERE id = ?;";
-        try(
-                Connection connection = dataSource.getConnection();
-                PreparedStatement preparedStatement = connection.prepareStatement(query)
-        ) {
+        String query = "SELECT * FROM security_info INNER JOIN public.role r " +
+                "ON r.role_id = security_info.role_id WHERE id = ?;";
+        try (Connection connection = dataSource.getConnection();
+             PreparedStatement preparedStatement = connection.prepareStatement(query)) {
             preparedStatement.setInt(1, id);
             ResultSet resultSet = preparedStatement.executeQuery();
             if (resultSet.next()) {
@@ -59,21 +62,37 @@ public class SecurityRepository {
             e.printStackTrace();
         }
 
-        return securityInfo;
+        return Optional.ofNullable(securityInfo);
     }
 
-    public SecurityInfo createSecurityInfo(SecurityInfo newSecurityInfo, int userId) {
-        SecurityInfo securityInfo = null;
-        String query = "INSERT INTO security_info (id, username, password, email) VALUES (?, ?, ?, ?);";
-        try(
-                Connection connection = dataSource.getConnection();
-                PreparedStatement preparedStatement =
-                        connection.prepareStatement(query, Statement.RETURN_GENERATED_KEYS)
-        ) {
+    public Optional<SecurityInfo> createSecurityInfo(SecurityInfo newSecurityInfo, int userId)
+            throws NoSuchUserException, NotEnoughDataException, EmailOrUsernameIsAlreadyUsedException {
+        if (findById(userId).isPresent()) {
+            return Optional.empty();
+        }
+        if (newSecurityInfo.getUsername() == null ||
+                newSecurityInfo.getPassword() == null ||
+                newSecurityInfo.getEmail() == null) {
+            throw new NotEnoughDataException();
+        }
+        if (userRepository.findById(userId).isEmpty()) {
+            throw new NoSuchUserException();
+        }
+        if (userRepository.emailOrUsernameIsUsed(newSecurityInfo.getEmail(), newSecurityInfo.getUsername())) {
+            throw new EmailOrUsernameIsAlreadyUsedException();
+        }
+        Optional<SecurityInfo> securityInfo = Optional.empty();
+        String query = "INSERT INTO security_info (id, username, password, email, role_id) " +
+                "VALUES (?, ?, ?, ?, ?);";
+        try (Connection connection = dataSource.getConnection();
+             PreparedStatement preparedStatement =
+                     connection.prepareStatement(query, Statement.RETURN_GENERATED_KEYS)) {
             preparedStatement.setInt(1, userId);
             preparedStatement.setString(2, newSecurityInfo.getUsername());
-            preparedStatement.setString(3, newSecurityInfo.getPassword());
+            preparedStatement.setString(3, passwordEncoder.encode(newSecurityInfo.getPassword()));
             preparedStatement.setString(4, newSecurityInfo.getEmail());
+            Role role = newSecurityInfo.getRole();
+            preparedStatement.setInt(5, role == Role.ADMIN ? 1 : 2);
 
             preparedStatement.executeUpdate();
             ResultSet res = preparedStatement.getGeneratedKeys();
@@ -87,32 +106,73 @@ public class SecurityRepository {
         return securityInfo;
     }
 
-    public void delete(int userId) {
+    @Transactional
+    public void delete(int userId) throws NoSuchUserException {
+        if (userRepository.findById(userId).isEmpty()) {
+            throw new NoSuchUserException();
+        }
         String query = "DELETE FROM security_info WHERE id = ?;";
-        try(
-                Connection connection = dataSource.getConnection();
-                PreparedStatement preparedStatement =
-                        connection.prepareStatement(query, Statement.RETURN_GENERATED_KEYS)
-        ) {
-            preparedStatement.setInt(1, userId);
-            preparedStatement.executeUpdate();
+        String queryDeleteFromUserTable = "DELETE FROM stocks_user WHERE id = ?;";
+        try (Connection connection = dataSource.getConnection();
+             PreparedStatement preparedStatement =
+                     connection.prepareStatement(query, Statement.RETURN_GENERATED_KEYS);
+             PreparedStatement preparedStatement2 =
+                     connection.prepareStatement(queryDeleteFromUserTable, Statement.RETURN_GENERATED_KEYS))
+        {
+            try {
+                connection.setAutoCommit(false);
+                preparedStatement.setInt(1, userId);
+                preparedStatement2.setInt(1, userId);
+
+                preparedStatement.executeUpdate();
+                preparedStatement2.executeUpdate();
+                connection.commit();
+            } catch (SQLException e) {
+                connection.rollback();
+                e.printStackTrace();
+            }
         } catch (SQLException e) {
             e.printStackTrace();
         }
     }
 
-    public SecurityInfo update(SecurityInfo updatedInfo, int userId) {
-        SecurityInfo securityInfo = null;
-        String query = "UPDATE security_info SET email = ?, password = ?, username = ? WHERE id = ?;";
-        try(
-                Connection connection = dataSource.getConnection();
-                PreparedStatement preparedStatement =
-                        connection.prepareStatement(query, Statement.RETURN_GENERATED_KEYS)
+    public Optional<SecurityInfo> update(SecurityInfo updatedInfo, int userId)
+            throws NoSuchUserException, EmailOrUsernameIsAlreadyUsedException {
+        if (userRepository.findById(userId).isEmpty()) {
+            throw new NoSuchUserException();
+        }
+        SecurityInfo oldSecurityInfo = findById(userId).get();
+        if (updatedInfo.getPassword() == null) {
+            updatedInfo.setPassword(oldSecurityInfo.getPassword());
+        }
+        else {
+            updatedInfo.setPassword(passwordEncoder.encode(updatedInfo.getPassword()));
+        }
+        if (updatedInfo.getEmail() == null) {
+            updatedInfo.setEmail(oldSecurityInfo.getEmail());
+        }
+        if (updatedInfo.getUsername() == null) {
+            updatedInfo.setUsername(oldSecurityInfo.getUsername());
+        }
+        if (updatedInfo.getRole() == null) {
+            updatedInfo.setRole(oldSecurityInfo.getRole());
+        }
+        if (userRepository.emailOrUsernameIsUsed(updatedInfo.getEmail(), updatedInfo.getUsername())) {
+            throw new EmailOrUsernameIsAlreadyUsedException();
+        }
+        Optional<SecurityInfo> securityInfo = Optional.empty();
+        String query = "UPDATE security_info SET email = ?, password = ?, username = ?, role_id = ?" +
+                " WHERE id = ?;";
+        try (Connection connection = dataSource.getConnection();
+             PreparedStatement preparedStatement =
+                     connection.prepareStatement(query, Statement.RETURN_GENERATED_KEYS)
         ) {
             preparedStatement.setString(1, updatedInfo.getEmail());
             preparedStatement.setString(2, updatedInfo.getPassword());
             preparedStatement.setString(3, updatedInfo.getUsername());
-            preparedStatement.setInt(4, userId);
+            Role role = updatedInfo.getRole();
+            preparedStatement.setInt(4, role == Role.ADMIN ? 1 : 2);
+            preparedStatement.setInt(5, userId);
 
             preparedStatement.executeUpdate();
             ResultSet res = preparedStatement.getGeneratedKeys();
